First rework of setup_package()
Doing this recursively proved to be quite challenging, especially setting templates. Hence, for now it is more of a workaround than a proper alternative. However it allows users to set up to five authors with setup package and redirects them to the new add_author function if they want to add an additional one. This function now takes vectors of ORCID numbers or manual entries as inputs. This greatly increases speed of creation of the description file.